+++ app/src/main/java/com/example/hope_uj_clinic/my_flow/ConditionData.java
@@ -2,47 +2,47 @@ package com.example.hope_uj_clinic.my_flow;
 
 public class ConditionData {
 
-    private int HR;
-    private int RESP;
-    private int SpO2;
-    private int TEMP;
+    private double HR;
+    private double RESP;
+    private double SpO2;
+    private double TEMP;
 
-    public ConditionData(int HR, int RESP, int spO2, int TEMP) {
+    public ConditionData(double HR, double RESP, double spO2, double TEMP) {
         this.HR = HR;
         this.RESP = RESP;
         SpO2 = spO2;
         this.TEMP = TEMP;
     }
 
-    public int getHR() {
+    public double getHR() {
         return HR;
     }
 
-    public void setHR(int HR) {
+    public void setHR(double HR) {
         this.HR = HR;
     }
 
-    public int getRESP() {
+    public double getRESP() {
         return RESP;
     }
 
-    public void setRESP(int RESP) {
+    public void setRESP(double RESP) {
         this.RESP = RESP;
     }
 
-    public int getSpO2() {
+    public double getSpO2() {
         return SpO2;
     }
 
-    public void setSpO2(int spO2) {
+    public void setSpO2(double spO2) {
         SpO2 = spO2;
     }
 
-    public int getTEMP() {
+    public double getTEMP() {
         return TEMP;
     }
 
-    public void setTEMP(int TEMP) {
+    public void setTEMP(double TEMP) {
         this.TEMP = TEMP;
     }
